+++ src/main/java/roomescape/controller/LoginController.java
@@ -7,10 +7,10 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.LoginCheckResponse;
-import roomescape.dto.LoginMember;
-import roomescape.dto.LoginRequest;
-import roomescape.dto.TokenDto;
+import roomescape.dto.login.LoginCheckResponse;
+import roomescape.dto.login.LoginMember;
+import roomescape.dto.login.LoginRequest;
+import roomescape.dto.token.TokenDto;
 import roomescape.service.AuthService;
 
 @RestController
