+++ src/main/java/roomescape/controller/MemberController.java
@@ -4,7 +4,7 @@ import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.memberDto.MemberResponse;
+import roomescape.dto.member.MemberResponse;
 import roomescape.service.MemberService;
 
 @RestController
