+++ src/main/java/roomescape/dto/login/LoginCheckResponse.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.dto.login;
 
 public record LoginCheckResponse(
         String name
