+++ src/main/java/roomescape/dto/reservation/UserReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto.reservationDto;
+package roomescape.dto.reservation;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDate;
