+++ src/main/java/roomescape/dto/reservationtime/ReservationTimeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.dto.reservationtime;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
