+++ src/main/java/roomescape/dto/token/TokenDto.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.dto.token;
 
 public record TokenDto(
         String accessToken
