+++ src/main/java/roomescape/infrastructure/JwtProvider.java
@@ -3,7 +3,7 @@ package roomescape.infrastructure;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
 import org.springframework.stereotype.Component;
-import roomescape.dto.MemberPayload;
+import roomescape.dto.member.MemberPayload;
 
 @Component
 public class JwtProvider {
@@ -11,6 +11,7 @@ public class JwtProvider {
     private final String secretKey = "Yn2kjibddFAWtnPJ2AFlL8WXmohJMCvigQggaEypa5E";
 
     public String createToken(MemberPayload memberPayload) {
+
         return Jwts.builder()
                 .setSubject(memberPayload.id())
                 .claim("name", memberPayload.name())
