+++ src/main/java/roomescape/service/ReservationService.java
@@ -8,8 +8,8 @@ import roomescape.domain.Reservation.Reservation;
 import roomescape.domain.ReservationTime.ReservationTime;
 import roomescape.domain.Theme.Theme;
 import roomescape.domain.member.Member;
-import roomescape.dto.reservationDto.ReservationRequest;
-import roomescape.dto.reservationDto.ReservationResponse;
+import roomescape.dto.reservation.ReservationRequest;
+import roomescape.dto.reservation.ReservationResponse;
 import roomescape.exception.ClientErrorExceptionWithLog;
 import roomescape.repository.MemberRepository;
 import roomescape.repository.ReservationRepository;
